+++ ARoQEngine/GameUtility/Base/Private/CharacterCode/Source/GUAscii.cpp
@@ -113,7 +113,7 @@ bool Ascii::FromUTF32(const uint32* input, const uint64 inputElementSize, uint8*
 *  @param[in] CharacterDecodeResult* result 変換結果
 *  @return    const tchar* 文字コード名
 *************************************************************************/
-bool Ascii::ToUTF16(const uint8* input, const uint64 inputByteSize, uint32* output, const uint64 outputElementSize, CharacterDecodeResult* result) const
+bool Ascii::ToUTF16(const uint8* input, const uint64 inputByteSize, uint16* output, const uint64 outputElementSize, CharacterDecodeResult* result) const
 {
 	/*-------------------------------------------------------------------
 	-              引数チェック
