+++ ARoQEngine/GameUtility/Base/Private/CharacterCode/Source/GUShiftJIS.cpp
@@ -8862,7 +8862,7 @@ bool ShiftJIS::FromUTF32(const uint32* input, const uint64 inputByteSize, uint8*
 *  @param[in] CharacterDecodeResult* result 変換結果
 *  @return    const tchar* 文字コード名
 *************************************************************************/
-bool ShiftJIS::ToUTF16(const uint8* input, const uint64 inputByteSize, uint32* output, const uint64 outputElementSize, CharacterDecodeResult* result) const
+bool ShiftJIS::ToUTF16(const uint8* input, const uint64 inputByteSize, uint16* output, const uint64 outputElementSize, CharacterDecodeResult* result) const
 {
 	/*-------------------------------------------------------------------
 	-              引数チェック
